// restitution-app/ClientApp/src/model/createVictimRestitutionCaseRequestDto.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Restitution API
 * API for the Restitution Application
 * OpenAPI spec version: v1
 */
import type { CourtInfoDto } from './courtInfoDto';
import type { DocumentDto } from './documentDto';
import type { ParticipantDto } from './participantDto';
import type { VictimApplicationDto } from './victimApplicationDto';

export interface CreateVictimRestitutionCaseRequestDto {
  /** @nullable */
  courtInfoCollection?: CourtInfoDto[] | null;
  /** @nullable */
  providerCollection?: ParticipantDto[] | null;
  /** @nullable */
  documentCollection?: DocumentDto[] | null;
  application?: VictimApplicationDto;
}