// restitution-app/ClientApp/src/model/victimApplicationDto.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Restitution API
 * API for the Restitution Application
 * OpenAPI spec version: v1
 */

export interface VictimApplicationDto {
  /** @nullable */
  middleName?: string | null;
  /** @nullable */
  otherFirstName?: string | null;
  /** @nullable */
  otherLastName?: string | null;
  /** @nullable */
  gender?: number | null;
  /** @nullable */
  indigenousStatus?: number | null;
  /** @nullable */
  preferredMethodOfContact?: number | null;
  /** @nullable */
  smsPreferred?: number | null;
  /** @nullable */
  primaryPhoneNumber?: string | null;
  /** @nullable */
  alternatePhoneNumber?: string | null;
  /** @nullable */
  email?: string | null;
  /** @minLength 1 */
  primaryAddressLine1: string;
  /** @nullable */
  primaryAddressLine2?: string | null;
  /** @nullable */
  primaryAddressLine3?: string | null;
  /** @minLength 1 */
  primaryCity: string;
  /** @minLength 1 */
  primaryProvince: string;
  /** @minLength 1 */
  primaryPostalCode: string;
  /** @minLength 1 */
  primaryCountry: string;
  /** @nullable */
  offenderFirstName?: string | null;
  /** @nullable */
  offenderMiddleName?: string | null;
  /** @nullable */
  offenderLastName?: string | null;
  /** @nullable */
  voicemailOption?: number | null;
  /** @minLength 1 */
  signature: string;
  /** @nullable */
  declarationFullName?: string | null;
  /** @nullable */
  signingOfficerTitle?: string | null;
  /** @nullable */
  declarationDate?: string | null;
  /** @nullable */
  contactTitle?: string | null;
  /** @nullable */
  offenderCustodyLocation?: string | null;
  /** @nullable */
  genderIdentityText?: string | null;
  /** @nullable */
  primaryRaceEthnicity?: number | null;
  /** @nullable */
  primaryRaceEthnicityText?: string | null;
  /** @nullable */
  pronouns?: number | null;
  /** @nullable */
  pronounsText?: string | null;
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  birthDate: string;
}